+++ src/pages/home/index.js
@@ -7,17 +7,14 @@ import Sample from "../../components/sampleComponent/sample.component";
 import RecentAcitivity  from "../../components/recentActivityComponent/recentActivity.component";
 import Layout from "../Layout";
 import { CardDeck } from "reactstrap";
+import RecentActivity from "../../components/recentActivityComponent/recentActivity.component";
 
 const Home = () => {
   return (
     <Layout>
       Home Page
       <Sample />
-      <CardDeck>
-        <RecentAcitivity />
-        <RecentAcitivity />
-        <RecentAcitivity />
-      </CardDeck>
+      <RecentActivity />
       
     </Layout>
   );
